refactor(state_dict): move into unified checkpoint
