+++ paddlenlp/trainer/plugins/unified_checkpoint.py
@@ -18,6 +18,7 @@ import json
 import multiprocessing
 import os
 
+import numpy as np
 import paddle
 import paddle.distributed as dist
 from paddle.distributed import fleet
@@ -27,7 +28,6 @@ from paddlenlp.trainer.trainer_utils import ExplicitEnum
 from paddlenlp.trainer.utils.helper import distributed_file, distributed_isfile
 from paddlenlp.transformers.model_utils import (
     PretrainedModel,
-    _load_state_dict_into_model,
     get_parameter_dtype,
     load_state_dict,
     unwrap_model,
@@ -208,8 +208,6 @@ def load_unified_checkpoint_locally(args, model, resume_from_checkpoint: str, sa
     if not isinstance(resolved_archive_file, list):
         resolved_archive_file = [resolved_archive_file]
 
-    error_msgs = []
-
     if len(resolved_archive_file) > 1:
         resolved_archive_file = tqdm(resolved_archive_file, desc="Loading checkpoint shards")
 
@@ -237,19 +235,11 @@ def load_unified_checkpoint_locally(args, model, resume_from_checkpoint: str, sa
                 None, model.config, state_dict=state_dict, ignore_error=len(resolved_archive_file) > 1
             )
 
-        error_msgs += _load_state_dict_into_model(model, state_dict, "")
+        unified_checkpoint_load_state_dict_into_model(model, state_dict, "")
 
         # force memory release
         del state_dict
         gc.collect()
-
-    if len(error_msgs) > 0:
-        error_msg = "\n\t".join(error_msgs)
-        if " but the expected shape is" in error_msg:
-            error_msg += (
-                "\n\tYou may consider adding `ignore_mismatched_sizes=True` in the model `from_pretrained` method."
-            )
-        raise RuntimeError(f"Error(s) in loading state_dict for {model.__class__.__name__}:\n\t{error_msg}")
 
 
 def unified_checkpoint_into_shards(
@@ -914,10 +904,7 @@ def load_unified_checkpoint_dynamically(args, model, optimizer, resume_from_chec
     )
     dist.barrier()
 
-    error_msgs = _load_state_dict_into_model(model, state_dict, "")
-    if len(error_msgs) > 0:
-        error_msg = "\n\t".join(error_msgs)
-        raise RuntimeError(f"Error(s) in loading dynamic state_dict for {model.__class__.__name__}:\n\t{error_msg}")
+    unified_checkpoint_load_state_dict_into_model(model, state_dict, "")
 
 
 def load_unified_optimizer_dynamically(args, model, optimizer, resume_from_checkpoint, safe_serialization=False):
@@ -1671,3 +1658,39 @@ def is_need_master_weight(optimizer, is_fp16_or_bp16):
         return optimizer._multi_precision and is_fp16_or_bp16
     else:
         return False
+
+
+def unified_checkpoint_load_state_dict_into_model(model_to_load, state_dict, start_prefix):
+    def is_0d_or_1d(tensor):
+        return len(tensor.shape) == 0 or list(tensor.shape) == [1]
+
+    if len(start_prefix) > 0:
+        for key in list(state_dict.keys()):
+            if key.startswith(start_prefix):
+                state_dict[key.replace(start_prefix, "")] = state_dict.pop(key)
+
+    expected_place = paddle.framework._current_expected_place()
+    for key, value in model_to_load.state_dict().items():
+        if key in state_dict:
+            value_state_dict = state_dict.pop(key)
+            if isinstance(value_state_dict, np.ndarray):
+                raise ValueError(
+                    "convert_state_dict_dtype expected paddle.Tensor not numpy.ndarray, plase convert numpy.ndarray to paddle.Tensor"
+                )
+            # confirm parameter cast is executed on the same device as model
+            # Note: cast(FP32 -> FP16) has diff on different devices, need to fix it
+            if value_state_dict.is_floating_point() and value_state_dict.dtype != value.dtype:
+                if value_state_dict.place._equal(expected_place):
+                    value_state_dict = paddle.cast(value_state_dict, value.dtype)
+                else:
+                    # confirm buffer cast is executed on the same device as model
+                    value_state_dict = paddle.cast(value_state_dict._copy_to(expected_place, False), value.dtype)
+
+            # unified 0d and 1d tensor
+            if is_0d_or_1d(value) and is_0d_or_1d(value_state_dict):
+                if list(value.shape) != list(value_state_dict.shape):
+                    value_state_dict = paddle.reshape(value_state_dict, value.shape)
+            value.set_value(value_state_dict)
+            # del value_state_dict
+
+    del state_dict
