+++ paddlenlp/transformers/model_utils.py
@@ -708,41 +708,26 @@ def faster_set_state_dict(model, state_dict):
 
 
 def _load_state_dict_into_model(model_to_load, state_dict, start_prefix):
-    def is_0d_or_1d(tensor):
-        return len(tensor.shape) == 0 or list(tensor.shape) == [1]
+    # torch will cast dtype in load_state_dict, but paddle strictly check dtype
+    _convert_state_dict_dtype_and_shape(state_dict, model_to_load)
 
+    error_msgs = []
     if len(start_prefix) > 0:
         for key in list(state_dict.keys()):
             if key.startswith(start_prefix):
                 state_dict[key.replace(start_prefix, "")] = state_dict.pop(key)
 
-    expected_place = paddle.framework._current_expected_place()
-    for key, value in model_to_load.state_dict().items():
-        if key in state_dict:
-            value_state_dict = state_dict.pop(key)
-            if isinstance(value_state_dict, np.ndarray):
-                raise ValueError(
-                    "convert_state_dict_dtype expected paddle.Tensor not numpy.ndarray, plase convert numpy.ndarray to paddle.Tensor"
-                )
-            # confirm parameter cast is executed on the same device as model
-            # Note: cast(FP32 -> FP16) has diff on different devices, need to fix it
-            if value_state_dict.is_floating_point() and value_state_dict.dtype != value.dtype:
-                if value_state_dict.place._equal(expected_place):
-                    value_state_dict = paddle.cast(value_state_dict, value.dtype)
-                else:
-                    # confirm buffer cast is executed on the same device as model
-                    value_state_dict = paddle.cast(value_state_dict._copy_to(expected_place, False), value.dtype)
-
-            # unified 0d and 1d tensor
-            if is_0d_or_1d(value) and is_0d_or_1d(value_state_dict):
-                if list(value.shape) != list(value_state_dict.shape):
-                    value_state_dict = paddle.reshape(value_state_dict, value.shape)
-            value.set_value(value_state_dict)
-            # del value_state_dict
+    # TODO: add return status to state_dict
+    with warnings.catch_warnings(record=True) as w:
+        warnings.resetwarnings()
+        # paddlenlp hold  missing_keys , just ignore not found warnings.
+        warnings.filterwarnings("ignore", message=r".*is not found in the provided dict.*")
+        model_to_load.set_state_dict(state_dict)
+        error_msgs.extend([str(x.message) for x in w])
 
     del state_dict
 
-    return []
+    return error_msgs
 
 
 def _convert_state_dict_dtype_and_shape(state_dict, model_to_load):
